+++ kernel/include/nano_internal.h
@@ -16,7 +16,6 @@
 
 #include <kernel.h>
 #include <kernel_structs.h>
-#include <ksched.h>
 
 #ifndef _ASMLANGUAGE
 
@@ -51,65 +50,6 @@ extern void _setup_new_thread(struct k_thread *new_thread,
 			      k_thread_entry_t entry,
 			      void *p1, void *p2, void *p3,
 			      int prio, u32_t options);
-
-#ifdef CONFIG_TIMESLICING
-extern void _update_time_slice_before_swap(void);
-#else
-#define _update_time_slice_before_swap() /**/
-#endif
-
-#ifdef CONFIG_STACK_SENTINEL
-extern void _check_stack_sentinel(void);
-#else
-#define _check_stack_sentinel() /**/
-#endif
-
-/* context switching and scheduling-related routines */
-#ifdef CONFIG_USE_SWITCH
-
-/* New style context switching.  _arch_switch() is a lower level
- * primitive that doesn't know about the scheduler or return value.
- * Needed for SMP, where the scheduler requires spinlocking that we
- * don't want to have to do in per-architecture assembly.
- */
-static inline unsigned int _Swap(unsigned int key)
-{
-	struct k_thread *new_thread, *old_thread;
-	int ret;
-
-	old_thread = _kernel.current;
-
-	_check_stack_sentinel();
-	_update_time_slice_before_swap();
-
-	new_thread = _get_next_ready_thread();
-
-	old_thread->swap_retval = -EAGAIN;
-
-	_kernel.current = new_thread;
-	_arch_switch(new_thread->switch_handle,
-		     &old_thread->switch_handle);
-
-	ret =_kernel.current->swap_retval;
-
-	irq_unlock(key);
-
-	return ret;
-}
-
-#else /* !CONFIG_USE_SWITCH */
-
-extern unsigned int __swap(unsigned int key);
-
-static inline unsigned int _Swap(unsigned int key)
-{
-	_check_stack_sentinel();
-	_update_time_slice_before_swap();
-
-	return __swap(key);
-}
-#endif
-
 
 #ifdef CONFIG_USERSPACE
 /**
