+++ kernel/init.c
@@ -28,6 +28,7 @@
 #include <string.h>
 #include <misc/dlist.h>
 #include <nano_internal.h>
+#include <kswap.h>
 
 /* kernel build timestamp items */
