+++ kernel/mailbox.c
@@ -17,7 +17,7 @@
 #include <wait_q.h>
 #include <misc/dlist.h>
 #include <init.h>
-#include <nano_internal.h>
+#include <kswap.h>
 
 #if (CONFIG_NUM_MBOX_ASYNC_MSGS > 0)
