+++ kernel/thread.c
@@ -25,6 +25,7 @@
 #include <atomic.h>
 #include <syscall_handler.h>
 #include <nano_internal.h>
+#include <kswap.h>
 
 extern struct _static_thread_data _static_thread_data_list_start[];
 extern struct _static_thread_data _static_thread_data_list_end[];
