+++ kernel/thread_abort.c
@@ -13,6 +13,7 @@
 #include <kernel.h>
 #include <kernel_structs.h>
 #include <nano_internal.h>
+#include <kswap.h>
 #include <string.h>
 #include <toolchain.h>
 #include <linker/sections.h>
